+++ packages/jugg/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - [ ] refactor: 集中式管理 webpack 插件、loader 等等的配置项，并对外提供获取、修改的途径
 - feat: 新增 `jugg inspect` 命令用于查看配置
-- [ ] feat: 命令行支持 `mode=` 以配置 `NODE_ENV`，默认 `development`
+- feat: 命令行支持 `mode=` 以配置 `NODE_ENV`，默认 `development`
 
 ## 0.2.0 - 2019-11-11